requirements.txt
Click==7.0
Flask==1.1.1
Flask-MySQL==1.4.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
PyMySQL==0.9.3
Werkzeug==0.16.0